% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccat_reporting_utils.R
\name{iccat_fill_report_dataset_characteristics}
\alias{iccat_fill_report_dataset_characteristics}
\title{Fills an ICCAT template workbook with dataset characteristics information}
\usage{
iccat_fill_report_dataset_characteristics(
 wb, reporting_flag, from, to,
 report_version, report_type, report_coverage, notes
)
}
\arguments{
\item{wb}{a \code{Workbook} object from \pkg{openxlsx}}

\item{reporting_flag}{reporting flag}

\item{from}{start year/date}

\item{to}{end year/date}

\item{report_version}{report version (\code{Final}, \code{Preliminary})}

\item{report_type}{report type}

\item{report_coverage}{report coverage}

\item{notes}{notes}
}
\value{
the modified \code{Workbook} object
}
\description{
Fills an ICCAT template workbook with dataset characteristics information
}
