% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccat_reporting_utils.R
\name{iccat_fill_report_statistical_correspondent}
\alias{iccat_fill_report_statistical_correspondent}
\title{Fills an ICCAT template workbook with statistical correspondent information}
\usage{
iccat_fill_report_statistical_correspondent(wb, metadata)
}
\arguments{
\item{wb}{a \code{Workbook} object from \pkg{openxlsx}}

\item{metadata}{a named \link{list} with business metadata}
}
\value{
the modified \code{Workbook} object
}
\description{
Fills an ICCAT template workbook with statistical correspondent information
}
