% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfishr.R
\docType{package}
\name{repfishr}
\alias{repfishr-package}
\alias{repfishr}
\title{Fisheries data reporting flows}
\description{
R framework to facilitate fisheries data reporting flows, with first intention to help countries to validate, 
format and send data to relevant organizations including regional fisheries management organizations or bodies handling #' data collection reference frameworks, and FAO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdiwg/repfishr}
  \item \url{https://fdiwg.github.io/repfishr/}
  \item Report bugs at \url{https://github.com/fdiwg/repfishr/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
