% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_actor.R
\docType{class}
\name{reporting_actor}
\alias{reporting_actor}
\title{reporting_actor}
\format{
\code{\link[R6]{R6Class}}
}
\value{
Object of class \code{\link[R6]{R6Class}} for modelling a reporting actor
}
\description{
reporting_actor

reporting_actor
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{actor}
\keyword{reporting}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name,}

\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reporting_actor-new}{\code{reporting_actor$new()}}
\item \href{#method-reporting_actor-clone}{\code{reporting_actor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_actor-new"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_actor-new}{}}}
\subsection{Method \code{new()}}{
Initializes a reporting sender
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_actor$new(id, name = NULL, type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_actor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_actor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_actor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
