% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_flow.R
\docType{class}
\name{reporting_flow}
\alias{reporting_flow}
\title{reporting_flow}
\format{
\code{\link[R6]{R6Class}}
}
\value{
Object of class \code{\link[R6]{R6Class}} for modelling a reporting flow
}
\description{
reporting_flow

reporting_flow
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{flow}
\keyword{reporting}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sender}}{sender object of class \link{reporting_sender}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reporting_flow-new}{\code{reporting_flow$new()}}
\item \href{#method-reporting_flow-getReceivers}{\code{reporting_flow$getReceivers()}}
\item \href{#method-reporting_flow-getReceiverIds}{\code{reporting_flow$getReceiverIds()}}
\item \href{#method-reporting_flow-getReceiver}{\code{reporting_flow$getReceiver()}}
\item \href{#method-reporting_flow-clone}{\code{reporting_flow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_flow-new"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_flow-new}{}}}
\subsection{Method \code{new()}}{
Initializes a reporting flow
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_flow$new(sender, sender_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sender}}{sender id}

\item{\code{sender_type}}{type of sender}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_flow-getReceivers"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_flow-getReceivers}{}}}
\subsection{Method \code{getReceivers()}}{
Get sender


Get list of valid receivers for the selected sender
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_flow$getReceivers(raw = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw}}{raw}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \link{reporting_sender}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_flow-getReceiverIds"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_flow-getReceiverIds}{}}}
\subsection{Method \code{getReceiverIds()}}{
Get list of valid receivers IDs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_flow$getReceiverIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_flow-getReceiver"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_flow-getReceiver}{}}}
\subsection{Method \code{getReceiver()}}{
Get a receiver by its id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_flow$getReceiver(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \link{reporting_organization}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_flow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_flow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_flow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
