% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_format.R
\docType{class}
\name{reporting_format}
\alias{reporting_format}
\title{reporting_format}
\format{
\code{\link[R6]{R6Class}}
}
\value{
Object of class \code{\link[R6]{R6Class}} for modelling a reporting format
}
\description{
reporting_format

reporting_format
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{format}
\keyword{reporting}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{ref}}{ref format specification link}

\item{\code{spec}}{format specification object of class \link[vrule]{format_spec}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reporting_format-new}{\code{reporting_format$new()}}
\item \href{#method-reporting_format-clone}{\code{reporting_format$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_format-new"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_format-new}{}}}
\subsection{Method \code{new()}}{
Initializes a reporting task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_format$new(id, name, ref)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{ref}}{ref}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_format-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_format-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_format$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
