% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_receiver.R
\docType{class}
\name{reporting_receiver}
\alias{reporting_receiver}
\title{reporting_receiver}
\format{
\code{\link[R6]{R6Class}}
}
\value{
Object of class \code{\link[R6]{R6Class}} for modelling a reporting receiver
}
\description{
reporting_receiver

reporting_receiver
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{receiver}
\keyword{reporting}
\section{Super class}{
\code{\link[repfishr:reporting_actor]{repfishr::reporting_actor}} -> \code{reporting_receiver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reporting_receiver-new}{\code{reporting_receiver$new()}}
\item \href{#method-reporting_receiver-clone}{\code{reporting_receiver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-new"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-new}{}}}
\subsection{Method \code{new()}}{
Initializes a reporting receiver
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$new(id, name, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
