% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_task.R
\docType{class}
\name{reporting_task}
\alias{reporting_task}
\title{reporting_task}
\format{
\code{\link[R6]{R6Class}}
}
\value{
Object of class \code{\link[R6]{R6Class}} for modelling a reporting task
}
\description{
reporting_task

reporting_task
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{reporting}
\keyword{task}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sender}}{sender}

\item{\code{receiver}}{receiver}

\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{context}}{context}

\item{\code{measurement}}{measurement}

\item{\code{formats}}{formats}

\item{\code{process_fun}}{process handler (function)}

\item{\code{report_fun}}{report handler (function)}

\item{\code{report_data}}{data to report}

\item{\code{report_metadata}}{metadata to report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reporting_task-new}{\code{reporting_task$new()}}
\item \href{#method-reporting_task-setSender}{\code{reporting_task$setSender()}}
\item \href{#method-reporting_task-process}{\code{reporting_task$process()}}
\item \href{#method-reporting_task-report}{\code{reporting_task$report()}}
\item \href{#method-reporting_task-clone}{\code{reporting_task$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_task-new"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_task-new}{}}}
\subsection{Method \code{new()}}{
Initializes a reporting task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_task$new(receiver = NULL, file = NULL, task = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{receiver}}{receiver}

\item{\code{file}}{file}

\item{\code{task}}{task (as list object)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_task-setSender"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_task-setSender}{}}}
\subsection{Method \code{setSender()}}{
Set sender
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_task$setSender(sender)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sender}}{sender object of class \link{reporting_sender}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_task-process"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_task-process}{}}}
\subsection{Method \code{process()}}{
Process data before reporting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_task$process(
  data,
  metadata,
  params = list(),
  path,
  parallel = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{object of class \link{data.frame}}

\item{\code{metadata}}{metadata object}

\item{\code{params}}{additional parameters. Optional. Default is an empty list}

\item{\code{path}}{path for the output file}

\item{\code{parallel}}{whether data validation should be run in parallel}

\item{\code{...}}{any other arguments to be passed to \pkg{vrule} validation method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_task-report"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_task-report}{}}}
\subsection{Method \code{report()}}{
Reports data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_task$report(
  data,
  metadata,
  params = list(),
  path,
  parallel = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{object of class \link{data.frame}}

\item{\code{metadata}}{metadata object}

\item{\code{params}}{additional parameters. Optional. Default is an empty list}

\item{\code{path}}{path for the output file}

\item{\code{parallel}}{whether data validation should be run in parallel}

\item{\code{...}}{any other arguments to be passed to \pkg{vrule} validation method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_task-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_task-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_task$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
