% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccat_reporting_utils.R
\name{iccat_fill_report_dataset}
\alias{iccat_fill_report_dataset}
\title{Fills an ICCAT template workbook with a dataset}
\usage{
iccat_fill_report_statistical_correspondent(wb, metadata)
}
\arguments{
\item{wb}{a \code{Workbook} object from \link{openxlsx}}

\item{sheet}{a sheet index}

\item{x}{an object of class \link{data.frame}}

\item{startRow}{start row index}

\item{startCol}{start column index}
}
\value{
the modified \code{Workbook} object
}
\description{
Fills an ICCAT template workbook with a dataset
}
