% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccat_reporting_utils.R
\name{iccat_fill_report_dataset_characteristics}
\alias{iccat_fill_report_dataset_characteristics}
\title{Fills an ICCAT template workbook with dataset characteristics information}
\usage{
iccat_fill_report_dataset_characteristics(
 wb, reporting_flag, from, to,
 report_version, report_type, report_coverage
)
}
\arguments{
\item{wb}{a \code{Workbook} object from \link{openxlsx}}

\item{metadata}{a named \link{list} with business metadata}
}
\value{
the modified \code{Workbook} object
}
\description{
Fills an ICCAT template workbook with dataset characteristics information
}
