% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_receiver.R
\docType{class}
\name{reporting_receiver}
\alias{reporting_receiver}
\title{reporting_receiver}
\format{
\code{\link[R6]{R6Class}}
}
\value{
Object of class \code{\link[R6]{R6Class}} for modelling a reporting receiver
}
\description{
reporting_receiver

reporting_receiver
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{receiver}
\keyword{reporting}
\section{Super class}{
\code{\link[repfishr:reporting_actor]{repfishr::reporting_actor}} -> \code{reporting_receiver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sender}}{sender}

\item{\code{tasks}}{tasks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reporting_receiver-new}{\code{reporting_receiver$new()}}
\item \href{#method-reporting_receiver-setSender}{\code{reporting_receiver$setSender()}}
\item \href{#method-reporting_receiver-getTasks}{\code{reporting_receiver$getTasks()}}
\item \href{#method-reporting_receiver-getTaskDefinitions}{\code{reporting_receiver$getTaskDefinitions()}}
\item \href{#method-reporting_receiver-getTaskDefinitionById}{\code{reporting_receiver$getTaskDefinitionById()}}
\item \href{#method-reporting_receiver-clone}{\code{reporting_receiver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-new"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-new}{}}}
\subsection{Method \code{new()}}{
Initializes a reporting receiver
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$new(id, name = NULL, type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-setSender"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-setSender}{}}}
\subsection{Method \code{setSender()}}{
Set sender
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$setSender(sender)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sender}}{sender object of class \link{reporting_sender}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-getTasks"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-getTasks}{}}}
\subsection{Method \code{getTasks()}}{
Get task IDs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$getTasks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the list of task IDs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-getTaskDefinitions"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-getTaskDefinitions}{}}}
\subsection{Method \code{getTaskDefinitions()}}{
Get tasks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$getTaskDefinitions(raw = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw}}{raw}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of \link{reporting_task}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-getTaskDefinitionById"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-getTaskDefinitionById}{}}}
\subsection{Method \code{getTaskDefinitionById()}}{
Get task definition by ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$getTaskDefinitionById(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \link{reporting_task}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reporting_receiver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reporting_receiver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reporting_receiver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
